+++ pagd/layouts/__init__.py
@@ -4,6 +4,9 @@
 # file 'LICENSE', which is part of this source code package.
 #       Copyright (c) 2013 R Pratap Chakravarthy
 
-"""Layout plugins to creat site-layouts and subsequently generate html files."""
+"""Layout plugins to creat site-layouts and subsequently generate html files.
+User documentation for ``pagd`` will cover most of the stuff that is dealt by
+layout-plugins.
+"""
 
 import pagd.layouts.myblog